Data.H3.Geo.Projection

Description

Scales for projecting geographic coordinates onto a 2D plane.

Projections

data Mercator a

mercator :: (Point Radians, Point Radians) -> NullMeridian -> ScaleOptions Mercator (Point Radians) (Double, Double)

Construct a mercator projection that maps the given "rectangle" of coordinates to the target rectangle

newtype NullMeridian

Null meridian used for mercator projection (in degrees)

Constructors

NullMeridian

Fields

getNullMeridian :: Degrees

greenwich :: NullMeridian

greenwich = NullMeridian 0

data Albers a

albers

Arguments

:: (Point Radians, Point Radians)
Area that should be mapped to the target area

-> Point Radians
Reference point

-> Radians
First standard parallel

-> Radians
Second standard parallel

-> ScaleOptions Albers (Point Radians) (Double, Double)

Albers projection (https:/en.wikipedia.orgwiki/Albers_projection). Sensible values for first and second standard parallels are 20°N50°N or 15°N45°N

albers'

Arguments

:: (Point Radians, Point Radians)
Area that should be mapped to the target area

-> Point Radians
Reference point

-> ScaleOptions Albers (Point Radians) (Double, Double)

Albers projection (https:/en.wikipedia.orgwiki/Albers_projection) using 15 and 45 degrees for the standard parallels.

Data constructors

data family ScaleOptions (f :: * -> *) a b :: *

Additional parameters for the scale (other than target range).